YARN-10412. Move CS placement rule related changes to a separate package. Contributed by Gergely Pollak
