+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/placement/MappingRuleCreator.java
@@ -27,11 +27,11 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.hadoop.yarn.server.resourcemanager.placement.MappingRule;
-import org.apache.hadoop.yarn.server.resourcemanager.placement.MappingRuleAction;
-import org.apache.hadoop.yarn.server.resourcemanager.placement.MappingRuleActions;
-import org.apache.hadoop.yarn.server.resourcemanager.placement.MappingRuleMatcher;
-import org.apache.hadoop.yarn.server.resourcemanager.placement.MappingRuleMatchers;
+import org.apache.hadoop.yarn.server.resourcemanager.placement.csmappingrule.MappingRule;
+import org.apache.hadoop.yarn.server.resourcemanager.placement.csmappingrule.MappingRuleAction;
+import org.apache.hadoop.yarn.server.resourcemanager.placement.csmappingrule.MappingRuleActions;
+import org.apache.hadoop.yarn.server.resourcemanager.placement.csmappingrule.MappingRuleMatcher;
+import org.apache.hadoop.yarn.server.resourcemanager.placement.csmappingrule.MappingRuleMatchers;
 
 // These are generated classes - use GeneratePojos class to create them
 // if they are missing
