+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/java/org/apache/hadoop/yarn/server/resourcemanager/placement/MockQueueHierarchyBuilder.java
@@ -36,7 +36,7 @@ import java.util.Set;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class MockQueueHierarchyBuilder {
+public class MockQueueHierarchyBuilder {
   private static final String ROOT = "root";
   private static final String QUEUE_SEP = ".";
   private List<String> queuePaths = Lists.newArrayList();
